refactor(compiler): move TransactionManager to own folder
